[IDLE-000] Refactor, 파일 구조 리팩토링
