+++ project/Projects/Presentation/Feature/Center/Sources/View/Setting/PasswordForDeregisterVC.swift
@@ -1,6 +1,6 @@
 //
 //  PasswordForDeregisterVC.swift
-//  RootFeature
+//  CenterFeature
 //
 //  Created by choijunios on 8/21/24.
 //
