+++ project/Projects/Presentation/Feature/Root/Sources/Screen/Common/Setting/Deregister/Coordinator/DeRegisterCoordinator.swift
@@ -7,25 +7,9 @@
 
 import UIKit
 import Entity
+import CenterFeature
 import PresentationCore
 import UseCaseInterface
-
-public protocol DeregisterCoordinatable: ParentCoordinator {
-    
-    /// 탈퇴과정이 끝났음을 알립니다.
-    func flowFinished()
-    
-    /// 공통: 탈퇴 이유를 선택합니다.
-    func showSelectReasonScreen()
-    
-    /// 센터관리자: 마지막으로 비밀번호를 입력합니다.
-    func showFinalPasswordScreen(reasons: [DeregisterReasonVO])
-    
-    /// 요양보호사: 마지막으로 전화번호를 입력합니다.
-    func showFinalPhoneAuthScreen(reasons: [DeregisterReasonVO])
-    
-    func coordinatorDidFinish()
-}
 
 public class DeRegisterCoordinator: DeregisterCoordinatable {
     
@@ -80,7 +64,7 @@ public class DeRegisterCoordinator: DeregisterCoordinatable {
     
     public func showFinalPasswordScreen(reasons: [Entity.DeregisterReasonVO]) {
     
-        let coordinator = FinalPasswordAuhCoordinator(
+        let coordinator = PasswordForDeregisterCoordinator(
             dependency: .init(
                 authUseCase: authUseCase,
                 reasons: reasons,
